backend/requirements.txt
alembic==1.14.0
amadeus==11.0.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
certifi==2024.12.14
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
Django==5.1.5
django-celery-beat==2.7.0
django-timezone-field==7.1
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.6
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
kombu==5.4.2
Mako==1.3.8
MarkupSafe==3.0.2
packaging==24.2
passlib==1.7.4
pluggy==1.5.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
pyasn1==0.6.1
pydantic==2.10.4
pydantic_core==2.27.2
pytest==8.3.4
pytest-asyncio==0.25.2
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-http-client==3.3.7
python-jose==3.3.0
redis==5.2.1
rsa==4.9
sendgrid==6.11.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlparse==0.5.3
starkbank-ecdsa==2.2.0
starlette==0.41.3
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.34.0
vine==5.1.0
wcwidth==0.2.13
